# api.yaml
openapi: 3.0.0
servers: []
info:
  description: This is a simple API
  version: "1.0.0"
  title: Simple Inventory API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
paths:
...
  /dasboard/performanceData:
    post:
      tags:
        - dasboard
      summary: data related to user performance
      operationId: performanceData
      description: sends data to be rendered in charts
      responses:
        "200":
          description: data for charts
          content:
            application/json:
              schema:
                type: object
                properties:
                  yearlyPerformance:
                    type: object
                    properties:
                      months:
                        type: array
                        example: ["Nov-2019", "Dec-2019", "Jan-2020"]
                      datasets:
                        $ref: "#/components/schemas/Datasets"
                  weeklyPerformance:
                    type: object
                    properties:
                      days:
                        type: array
                        example: ["Fri", "Sat", "Sun", "Mon"]
                      datasets:
                        $ref: "#/components/schemas/Datasets"
        "502":
          description: error processing request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: email@email.com
components:
  schemas:
    Datasets:
      type: array
      items:
        type: object
        properties:
          label:
            type: string
            example: Quizzes or Score or %Score
          data:
            type: array
            example: [12, 43, 6]
    Error:
      type: object
      properties:
        error:
          type: object
          required:
            - status
            - message
          properties:
            status:
              type: number
              example: 502
            message:
              type: string
              example: Email already registered